ENH: restructure into a package
Use flit as a build backend.
