+++ scpdt/_checker.py
